SpMV: zero y when the local block has no nonzeros
cuSPARSE 12 returns without writing y for a matrix with no nonzeros,
although beta is zero, so a rank whose local block has no columns got
stale values. Handle the empty block before calling the library.
The SpGEMM test tolerance goes back to 100 epsilon.

+++ Tests/Algebra/SpGEMM/main.cpp
@@ -259,7 +259,7 @@ int main (int argc, char *argv[])
             // Products are summed in different orders.
             AMREX_ALWAYS_ASSERT(all_true(amrex::almostEqual(ABT,BTAT,32)));
 
-            Real tol = std::numeric_limits<Real>::epsilon() * Real(1000);
+            Real tol = std::numeric_limits<Real>::epsilon() * Real(100);
             AMREX_ALWAYS_ASSERT(all_true(check_spmv(A, B, AB, pt3, tol)));
         }
